alz_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
nibabel>=5.2.0
SimpleITK>=2.3.1
torch>=2.3.0
torchvision>=0.18.0
monai>=1.3.0
pytest>=8.2.0
httpx>=0.27.0
openpyxl>=3.1.0
Pillow>=10.0.0
PyYAML>=6.0.2
